services/ingestion_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
google-cloud-storage==2.19.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.10.6